pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "relayx-device-sdk"
version = "0.1.0"
description = "RelayX Device SDK for Python — IoT device communication over NATS"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "nats-py>=2.9.0",
    "msgpack>=1.0.0",
    "ntplib>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
